+++ tests/unit/async_/io/test_neo4j_pool.py
@@ -23,7 +23,7 @@ from neo4j import (
     WRITE_ACCESS,
 )
 from neo4j._async.io import AsyncNeo4jPool
-from neo4j._io.deadline import Deadline
+from neo4j._deadline import Deadline
 from neo4j.addressing import ResolvedAddress
 from neo4j.conf import (
     PoolConfig,
