+++ tests/unit/mixed/io/test_direct.py
@@ -32,7 +32,7 @@ import time
 
 import pytest
 
-from neo4j._io.deadline import Deadline
+from neo4j._deadline import Deadline
 
 from ...async_.io.test_direct import AsyncFakeBoltPool
 from ...sync.io.test_direct import FakeBoltPool
