+++ tests/unit/sync/work/_fake_connection.py
@@ -21,7 +21,7 @@ import inspect
 import pytest
 
 from neo4j import ServerInfo
-from neo4j._io.deadline import Deadline
+from neo4j._deadline import Deadline
 from neo4j._sync.io import Bolt
